+++ database/migrations/2019_08_17_072409_create_hourly_wage_contract_histories_table.php
@@ -15,7 +15,7 @@ class CreateHourlyWageContractHistoriesTable extends Migration
     {
         Schema::create('hourly_wage_contract_histories', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('employee_id');
+            $table->bigInteger('employee_id');
             $table->integer('hourly_wage');
             $table->integer('over_time_hourly_extra_wage');
             $table->integer('midnight_hourly_extra_wage');
